+++ src/main/java/blusunrize/immersiveengineering/client/render/tile/MixerRenderer.java
@@ -9,20 +9,31 @@
 package blusunrize.immersiveengineering.client.render.tile;
 
 import blusunrize.immersiveengineering.client.utils.GuiHelper;
+import blusunrize.immersiveengineering.client.utils.IERenderTypes;
+import blusunrize.immersiveengineering.client.utils.RenderUtils;
 import blusunrize.immersiveengineering.common.blocks.metal.MixerBlockEntity;
 import blusunrize.immersiveengineering.common.register.IEBlocks.Multiblocks;
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.math.Matrix4f;
 import com.mojang.math.Quaternion;
 import com.mojang.math.Vector3f;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
+import net.minecraft.client.renderer.texture.OverlayTexture;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.client.extensions.common.IClientFluidTypeExtensions;
 import net.minecraftforge.client.model.data.ModelData;
 import net.minecraftforge.fluids.FluidStack;
+
+import static blusunrize.immersiveengineering.client.ClientUtils.getSprite;
+import static com.mojang.blaze3d.vertex.DefaultVertexFormat.*;
 
 public class MixerRenderer extends IEBlockEntityRenderer<MixerBlockEntity>
 {
@@ -60,8 +71,6 @@ public class MixerRenderer extends IEBlockEntityRenderer<MixerBlockEntity>
 		matrixStack.popPose();
 
 		matrixStack.translate(te.getFacing()==Direction.SOUTH||te.getFacing()==Direction.WEST?-.5: .5, -.625f, te.getFacing()==Direction.SOUTH||te.getFacing()==Direction.EAST?.5: -.5);
-		matrixStack.scale(.0625f, 1, .0625f);
-		matrixStack.mulPose(new Quaternion(90, 0, 0, true));
 
 		for(int i = te.tank.getFluidTypes()-1; i >= 0; i--)
 		{
@@ -69,10 +78,40 @@ public class MixerRenderer extends IEBlockEntityRenderer<MixerBlockEntity>
 			if(fs!=null&&fs.getFluid()!=null)
 			{
 				float yy = fs.getAmount()/(float)te.tank.getCapacity()*1.0625f;
-				matrixStack.translate(0, 0, -yy);
-				float w = (i < te.tank.getFluidTypes()-1||yy >= .125)?26: 16+yy/.0125f;
-				GuiHelper.drawRepeatedFluidSprite(bufferIn.getBuffer(RenderType.translucent()), matrixStack, fs,
-						-w/2, -w/2, w, w);
+				matrixStack.translate(0, yy, 0);
+				float w = (i < te.tank.getFluidTypes()-1||yy >= .125)?.8125f: 0.5f+yy*2.5f;
+				double px = w*16;
+
+				IClientFluidTypeExtensions props = IClientFluidTypeExtensions.of(fs.getFluid());
+				TextureAtlasSprite sprite = getSprite(props.getStillTexture(fs));
+				VertexConsumer consumer = bufferIn.getBuffer(RenderType.solid());
+				Matrix4f matrix4f = matrixStack.last().pose();
+
+				int col = props.getTintColor(fs);
+				float r = (col>>16&255)/255.0f;
+				float g = (col>>8&255)/255.0f;
+				float b = (col&255)/255.0f;
+
+				Vec3 from = new Vec3(-w, 0, -w);
+				Vec3 to = new Vec3(0, 0, 0);
+				for(int v = 0; v < 4; v++)
+				{
+					Vec3 start = from.add(v%2==1?w: 0, 0, v > 1?w: 0);
+					Vec3 end = to.add(v%2==1?w: 0, 0, v > 1?w: 0);
+					float uMin = sprite.getU(v%2==0?16-px: 0);
+					float uMax = sprite.getU(v%2==0?16: px);
+					float vMin = sprite.getV(v > 1?0: 16-px);
+					float vMax = sprite.getV(v > 1?px: 16);
+
+					consumer.vertex(matrix4f, (float)start.x, 0, (float)start.z).color(r, g, b, 1).uv(uMin, vMin)
+							.overlayCoords(OverlayTexture.NO_OVERLAY).uv2(combinedLightIn).normal(0, 1, 0).endVertex();
+					consumer.vertex(matrix4f, (float)start.x, 0, (float)end.z).color(r, g, b, 1).uv(uMin, vMax)
+							.overlayCoords(OverlayTexture.NO_OVERLAY).uv2(combinedLightIn).normal(0, 1, 0).endVertex();
+					consumer.vertex(matrix4f, (float)end.x, 0, (float)end.z).color(r, g, b, 1).uv(uMax, vMax)
+							.overlayCoords(OverlayTexture.NO_OVERLAY).uv2(combinedLightIn).normal(0, 1, 0).endVertex();
+					consumer.vertex(matrix4f, (float)end.x, 0, (float)start.z).color(r, g, b, 1).uv(uMax, vMin)
+							.overlayCoords(OverlayTexture.NO_OVERLAY).uv2(combinedLightIn).normal(0, 1, 0).endVertex();
+				}
 			}
 		}